+++ src/components/Button/Button.spec.js
@@ -3,12 +3,19 @@ import React from 'react';
 import Button from './Button';
 // import Test from './components/Test'
 
-const { enzyme: { mount }, chai: { expect }} = global;
+const {
+  enzyme: { mount },
+  chai: { expect },
+} = global;
 
 describe('<Button />', function() {
-  const wrapper = mount(<Button />);
+  const mounted = mount(<Button text="<Button /> spec test" />);
 
-  it('renders without errors', function() {
-    expect(wrapper).to.be.present();
+  it('mounts without errors', function() {
+    expect(mounted).to.be.present();
+  });
+
+  it('mounts with given text', function() {
+    expect(mounted).to.contain('<Button /> spec test');
   });
 });