+++ src/components/PropertyPicker/index.js
@@ -1,0 +1,3 @@
+import PropertyPicker from './PropertyPicker';
+
+export default PropertyPicker;