+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+# Changelog
+> generated with [@semantic-release/changelog](https://github.com/semantic-release/changelog)
+
+# [1.2.0](https://github.com/viewar/components/compare/v1.1.0...v1.2.0) (2019-11-16)
+
+
+### Bug Fixes
+
+* **Accordion:** fix propType for children ([175e38f](https://github.com/viewar/components/commit/175e38f3f6f192f0e622b1efd36c3220b84eec2f))
+* **Accordion:** fix propType for children ([0682674](https://github.com/viewar/components/commit/06826744b4469c22ef7312ac09ce2be762f48681))
+
+
+### Features
+
+* **spec:** add 'App.spec.js' ([4a7bc92](https://github.com/viewar/components/commit/4a7bc92c7fa4ed7bda30df27eefcc77dfe082e56))
+* **spec:** add 'App.spec.js' ([7037976](https://github.com/viewar/components/commit/703797652fe667cb2ddc68033030132c06d059de))
+* **viewar:** add 'viewar-core' and 'viewar-api' ([#12](https://github.com/viewar/components/issues/12)) ([25eb40a](https://github.com/viewar/components/commit/25eb40a7c4dc62e5a95048107ae7cb53774804e4))
+* **viewar:** add 'viewar-core' and 'viewar-api' ([#12](https://github.com/viewar/components/issues/12)) ([a964d86](https://github.com/viewar/components/commit/a964d8622086ab15702d5909ddfe4eb7dc54f12a))
+* **spec:** add spec test for `<Accordion />` ([d86962a](https://github.com/viewar/components/commit/d86962ac0b6b4b310fe8cd6630793dd1b4f2e963))
+* **spec:** add spec test for `<Accordion />` ([fa7edc2](https://github.com/viewar/components/commit/fa7edc279840698d24aa55f7254f3ec6850c49e1))
+
+# Changelog
+
+> generated with [@semantic-release/changelog](https://github.com/semantic-release/changelog)
+
+# [1.1.0](https://github.com/viewar/components/compare/v1.0.0...v1.1.0) (2019-11-15)
+
+### Features
+
+- **PropertyPicker:** add `<PropertyPicker />` ([a85ccf9](https://github.com/viewar/components/commit/a85ccf9c1bc261c5437d545d9102598b456d7656))
+
 # [1.0.0](https://github.com/viewar/components/compare/1a5ba55...v1.0.0) (2019-11-12)
 
 ### Bug Fixes